Restructure source code files
- Rename the ./lib directory to "src"
- Move `index.js` into the ./src directory
