+++ package.json
@@ -2,7 +2,7 @@
   "name": "gulp-gitstage",
   "version": "1.0.1",
   "description": "A plugin for Gulp to get an object stream of modified files on git.",
-  "main": "./index.js",
+  "main": "./src/index.js",
   "author": "Eric Cornelissen <ericornelissen@gmail.com>",
   "license": "MIT",
   "repository": {
